Add Collatz Conjecture module with tests
This commit introduces a new module for the Collatz Conjecture problem, which includes the main logic for calculating the number of steps required to reach 1 for a given number. In addition, a corresponding test suite has been added to validate the behavior of this module. It also includes an updated project file to reference the new module and its tests.