Add new exercise "Interest is Interesting"
This commit introduces a new exercise "Interest is Interesting". It includes an implementation of a savings account balance tracking program. It also involves the creation of tests to assert the accuracy of the calculations concerning interest rate, interest, annual balance update and the amount to donate to charities. A README detailing the exercise instructions and information about floating point numbers in F# is also included.