Add BirdWatcher exercise with related unit tests
This commit introduces the BirdWatcher exercise. It provides functionality to track and manipulate data related to bird sightings over a week. Additionally, fixtures for unit testing have been added to validate the behavior of the new functionality under various scenarios. A README file was also added to provide information and instructions. Minor changes were made to Grains.fs as well.

@coderabbitai ignore