Add F# implementation for the Yacht dice game
The Yacht game's functionality and related tests were implemented. This included creating the main Yacht module, YachtTests module, and the corresponding fsproj project file. Furthermore, a README file was added which provides instructions on how to play the game and score points. Lastly, a minor refactoring was performed in the TisburyTreasureHunt file.