Add Darts scoring application with unit tests
A new Darts application has been created to calculate earned points in a single toss of a Darts game. This script takes cartesian coordinates as input and returns the scoring based on the given rules. In addition, unit tests have been added to verify the functionality under different scenarios. The corresponding project file and README information are also included in this commit.