Added Allergies module with tests
This commit introduces a new Allergies module which includes the definition for different allergens and functions to calculate an allergy score. Additionally, the corresponding AllergiesTests module has been added to test the functionality of the new module. A README.md file with instructions for executing the tests has also been included.