Add RNA transcription functionality and tests
Three new files are added: RnaTranscription.fs, RnaTranscriptionTests.fs, and RnaTranscription.fsproj. This update introduces the functionality of transcribing DNA sequences into RNA sequences. Additionally, some tests for this functionality have been created and packed in the project file. Lastly, a README explaining all details about the newly introduced concept and how it is handled in the code is added.